spi_engine_execution: Add echoed SCLK support
There are boards (e.g. AD4624-30) which takes the SCLK and echos back to
th FPGA through a level shifter - doing this removes the effect of
round-trip timing delays from the level shifter. This is commonly done
whenever isolators are used since they are very slow.

By setting the ECHO_SCLK parameter to 1, the IP will use the incomming
echod SCLK clock to latch the SDI line(s). The sdi_data_valid is still
synchronous to the SPI clock, and it's generated after the last valid
SDI latch.

The designer responsibility to time the SDI shift registers in a wat
that respects the design requirments.